requirements_llm.txt
ragas
langchain-google-genai
langchain-google-vertexai
google-generativeai
datasets
